+++ connecteur-type/signature/SignatureRecuperation.php
@@ -30,6 +30,7 @@ class SignatureRecuperation extends ConnecteurTypeActionExecutor
         $document_orignal_element = $this->getMappingValue('document_original');
         $bordereau_element = $this->getMappingValue('bordereau');
         $annexe_element = $this->getMappingValue('autre_document_attache');
+        $multi_document_element = $this->getMappingValue('autre_document_a_signer');
         $multi_document_original_element = $this->getMappingValue('multi_document_original');
         $iparapheur_annexe_sortie_element = $this->getMappingValue('iparapheur_annexe_sortie');
         $iparapheur_metadata_sortie_element = $this->getMappingValue('iparapheur_metadata_sortie');
@@ -95,7 +96,8 @@ class SignatureRecuperation extends ConnecteurTypeActionExecutor
                     $document_element,
                     $document_orignal_element,
                     $multi_document_original_element,
-                    $annexe_element
+                    $annexe_element,
+                    $multi_document_element
                 );
             }
             return $return;
@@ -173,7 +175,8 @@ class SignatureRecuperation extends ConnecteurTypeActionExecutor
         $document_element,
         $document_orignal_element,
         $multi_document_original_element,
-        $annexe_element
+        $annexe_element,
+        $multi_document_element = ''
     ): bool {
 
         /** @var SignatureConnecteur $signature */
@@ -206,12 +209,19 @@ class SignatureRecuperation extends ConnecteurTypeActionExecutor
                 $donneesFormulaire->addFileFromData($document_orignal_element, $filename_orig, $document_original_data);
             }
 
-            if ($signature->hasMultiDocumentSigne($info)) {
+            $hasMultiDocumentSigne = $signature->hasMultiDocumentSigne($info);
+            $hasMultiDocumentFromStep = $multi_document_element
+                && $signature->supportsMultiDocument()
+                && $donneesFormulaire->get($multi_document_element);
+
+            if ($hasMultiDocumentSigne || $hasMultiDocumentFromStep) {
                 $this->addMultiDocumentSigne(
                     $signature->getAllDocumentSigne($info),
                     $multi_document_original_element,
                     $document_element,
-                    $annexe_element
+                    $annexe_element,
+                    $hasMultiDocumentFromStep ? $multi_document_element : '',
+                    $hasMultiDocumentSigne
                 );
             } else {
                 $this->addMainSignedFile($signature->getSignedFile($info), $document_element);
@@ -319,44 +329,94 @@ class SignatureRecuperation extends ConnecteurTypeActionExecutor
         array $allSignedFiles,
         string $multi_document_original_element,
         string $document_element,
-        string $annexe_element
+        string $annexe_element,
+        string $multi_document_element = '',
+        bool $annexesAreMultiDocuments = false
     ): void {
         $donneesFormulaire = $this->getDonneesFormulaire();
-        // Copie des $annexe_element dans $multi_document_original_element
-        if ($donneesFormulaire->get($annexe_element)) {
-            foreach ($donneesFormulaire->get($annexe_element) as $num => $fileName) {
-                $annexe_original_name = $donneesFormulaire->getFileName($annexe_element, $num);
-                $annexe_original_data = $donneesFormulaire->getFileContent($annexe_element, $num);
-                $filename = pathinfo($annexe_original_name, PATHINFO_FILENAME);
-                $extension = pathinfo($annexe_original_name, PATHINFO_EXTENSION);
-                $filename_orig = sprintf('%s_orig.%s', $filename, $extension);
-                $filename_orig = $this->getComputedFileName($filename_orig);
-                $donneesFormulaire->addFileFromData(
-                    $multi_document_original_element,
-                    $filename_orig,
-                    $annexe_original_data,
-                    $num
+
+        $multiDocumentBasenames = [];
+        if ($multi_document_element && $donneesFormulaire->get($multi_document_element)) {
+            foreach ($donneesFormulaire->get($multi_document_element) as $num => $fileName) {
+                $multiDocumentBasenames[] = pathinfo(
+                    $donneesFormulaire->getFileName($multi_document_element, $num),
+                    PATHINFO_FILENAME
                 );
             }
         }
 
-        $i = 0;
+        $originalIndex = $annexesAreMultiDocuments
+            ? $this->copyOriginalMultiDocuments($annexe_element, $multi_document_original_element, 0)
+            : 0;
+        $this->copyOriginalMultiDocuments(
+            $multi_document_element,
+            $multi_document_original_element,
+            $originalIndex
+        );
+
+        $originalFileBasename = pathinfo($donneesFormulaire->getFileName($document_element), PATHINFO_FILENAME);
+        $annexeIndex = 0;
+        $multiDocumentIndex = 0;
         foreach ($allSignedFiles as $file) {
             /** @var Fichier $file */
             $fileBasename = pathinfo($file->filename, PATHINFO_FILENAME);
-            $originalFileBasename = pathinfo($donneesFormulaire->getFileName($document_element), PATHINFO_FILENAME);
             if ($fileBasename === $originalFileBasename) {
                 $this->addMainSignedFile($file, $document_element);
+                continue;
+            }
+            if ($multi_document_element && in_array($fileBasename, $multiDocumentBasenames, true)) {
+                $donneesFormulaire->addFileFromData(
+                    $multi_document_element,
+                    $file->filename,
+                    $file->content,
+                    $multiDocumentIndex
+                );
+                $multiDocumentIndex++;
                 continue;
             }
             $donneesFormulaire->addFileFromData(
                 $annexe_element,
                 $file->filename,
                 $file->content,
-                $i
+                $annexeIndex
             );
-            $i++;
+            $annexeIndex++;
         }
+    }
+
+    /**
+     * @throws NotFoundException
+     * @throws Exception
+     * @return int l'index suivant disponible dans $multi_document_original_element
+     */
+    private function copyOriginalMultiDocuments(
+        string $source_element,
+        string $multi_document_original_element,
+        int $startIndex
+    ): int {
+        if (!$source_element) {
+            return $startIndex;
+        }
+        $donneesFormulaire = $this->getDonneesFormulaire();
+        if (!$donneesFormulaire->get($source_element)) {
+            return $startIndex;
+        }
+        foreach ($donneesFormulaire->get($source_element) as $num => $fileName) {
+            $original_name = $donneesFormulaire->getFileName($source_element, $num);
+            $original_data = $donneesFormulaire->getFileContent($source_element, $num);
+            $filename = pathinfo($original_name, PATHINFO_FILENAME);
+            $extension = pathinfo($original_name, PATHINFO_EXTENSION);
+            $filename_orig = sprintf('%s_orig.%s', $filename, $extension);
+            $filename_orig = $this->getComputedFileName($filename_orig);
+            $donneesFormulaire->addFileFromData(
+                $multi_document_original_element,
+                $filename_orig,
+                $original_data,
+                $startIndex
+            );
+            $startIndex++;
+        }
+        return $startIndex;
     }
 
     /**